Move executor into its own module
